Refactoring code Stats.cpp

+++ sfall/Modules/Stats.h
@@ -33,7 +33,7 @@ void _stdcall SetPCStatMax(int stat, int i);
 void _stdcall SetPCStatMin(int stat, int i);
 void _stdcall SetNPCStatMax(int stat, int i);
 void _stdcall SetNPCStatMin(int stat, int i);
-extern unsigned short standardApAcBonus;
-extern unsigned short extraApAcBonus;
+extern DWORD standardApAcBonus;
+extern DWORD extraApAcBonus;
 
 }
